dispatch/app/requirements.txt
# Python dependencies
PYTHON_VERSION=3.8

# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
python-dotenv==1.0.0

# OpenAI
openai==1.3.0

# Logging
logging==0.5.1.2

# CORS
starlette==0.27.0